requirements.txt
youtube_dl
discord.py